api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.7"
celery = {extras = ["redis"], version = "^5.3.4"}
djangorestframework = "^3.14.0"
django-filter = "^23.2"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
